Moved the majority of the external memory related structures and enumes to the auxiliary crate.
