+++ src/auxil/external-memory/Cargo.toml
@@ -12,3 +12,7 @@ edition = "2018"
 
 [lib]
 name = "external_memory"
+
+[dependencies]
+bitflags = "1.0"
+drm-fourcc = {version="2.1.1",features=["serde"]}
